feat:ㅇ 뒤로가기, 목록클릭시 시 원래 페이지로 가기, 새로고침, 정렬하기 포함      ㅇ 디폴트 링크 값 넣기      ㅇ 이미지 크게 하기      ㅇ 리스트 썸네일 크기 수정      ㅇ 스크랩 기능

+++ src/main/java/com/pickupluck/ecogging/domain/plogging/service/EventService.java
@@ -17,5 +17,5 @@ public interface EventService {
     Integer scrapEvent(Integer eventId)throws Exception;
     Integer updateView(Integer id) throws Exception;
     Boolean isEventScrap(Long userId, Integer eventId) throws Exception;
-    Boolean toggleEventScrap(Long userId, Integer eventId) throws Exception;
+    Boolean toggleEventScrap(Long userId, Long eventId) throws Exception;
 }
